finish comment ability

+++ app/Http/Controllers/CommentController.php
@@ -19,19 +19,39 @@ class CommentController extends Controller
         return redirect()->route('event', ['event_id' => $request->input('event_id')]);
     }
 
-    public function commentDeleteRequest(
-        $event_id,
-        $comment_id
-    ) {
-        $comment = Comment::find($comment_id);
+    public function commentDeleteRequest(Request $request)
+    {
+        if (Auth::user()->cannot('delete', Comment::find($request->input('comment_id')))) {
+            abort(403, 'Unauthorized action.');
+        }
+
+        $event_id = Comment::find($request->input('comment_id'))->event_id;
+        $comment = Comment::find($request->input('comment_id'));
         $comment->delete();
 
         return redirect()->route('event', ['event_id' => $event_id]);
     }
 
-    public function commentEditRequest(Request $request)
+    public function commentEdit($comment_id)
     {
-        $comment = Comment::find($request->input('event_id'));
+
+        if (Auth::user()->cannot('edit', Comment::find($comment_id))) {
+            abort(403, 'Unauthorized action.');
+        }
+
+        return view('events.comments.comment_edit', [
+            'comment' => Comment::find($comment_id),
+        ]);
+    }
+
+    public function commentEditRequest(Request $request, $comment_id)
+    {
+
+        if (Auth::user()->cannot('edit', Comment::find($comment_id))) {
+            abort(403, 'Unauthorized action.');
+        }
+
+        $comment = Comment::find($comment_id);
         $comment->comment = $request->input('comment');
         $comment->save();
